backend/requirements.txt
adrf==0.1.7
asgiref==3.8.1
async-property==0.2.2
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.41.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
requests==2.32.3
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
zope.interface==7.0.3
